Change the table of contents and apply requested changes
Equal titles on the same level as the previous section are skipped. Thus for Lorum.pdf this will not work, but I don't think any paper ever uses two equal subsequent titles on the same level.
Equal titles on a different level within the same section or with another section will work.